tests/test_auth.py
import pytest
from fastapi.testclient import TestClient
from app.main import app
from app.database import SessionLocal, init_db, Base, engine

client = TestClient(app)

@pytest.fixture(scope="module")
def test_db():
    Base.metadata.create_all(bind=engine)
    init_db()
    yield
    Base.metadata.drop_all(bind=engine)

def test_register_user(test_db):
    response = client.post("/auth/register/", json={"username": "testuser", "email": "testuser@example.com", "password": "testpassword"})
    assert response.status_code == 200
    assert response.json()["username"] == "testuser"

def test_login_user(test_db):
    response = client.post("/auth/login/", json={"username": "testuser", "password": "testpassword"})
    assert response.status_code == 200
    assert "access_token" in response.json()
